Moved FBProtoFSM to own file
